Save Tempfile instance in cache
We currently save the string *path* to tempfiles within the
`@temp_files` cache in the `Kubernetes` module. This is incorrect
because tempfiles are deleted when Ruby garbage collects the parent
`Tempfile` instance.

This change replaces the string path with the `Tempfile` instance so
that the files exist indefinitely. I also drop the function
`cache_temp_file` since it's unused and its behavior is incorrect with
regards to tempfiles.